+++ packages/react/src/components/Card/Card.spec.tsx
@@ -1,7 +1,6 @@
 import { fireEvent, render } from '@testing-library/react';
 import { byTestId, byText } from 'testing-library-selector';
 import Typography from '../../components/Typography';
-import { executeKeyPress } from '../../utils/test/executeKeyPress';
 import { axe } from '../../utils/test/axeConfig';
 import Card, { CardProps } from './Card';
 
@@ -55,15 +54,6 @@ describe('Card Component', () => {
 
     //Assert
     expect(ui.cardComp.get()).toBeInTheDocument();
-  });
-
-  it('selects card when Spacebar', () => {
-    const onClick = jest.fn();
-    // Act
-    renderComponent({ onClick });
-    executeKeyPress(ui.cardComp.get());
-    // Assert
-    expect(onClick).toBeCalled();
   });
 
   it(`doesn't renders right icon when hasRightIcon is false`, () => {