+++ packages/react/src/stories/Container/ContainerInfo.ts
@@ -1,0 +1,10 @@
+const TITLE = 'Container';
+const DESCRIPTION = `The 'Container' component allows us to add spacings as well as using it as a flexbox and adjusting direction and children gap.`;
+
+const DO = ['Use this component as a wrapper to your custom components'];
+
+const DONT = [
+  `Unneccessarily create a new component that can use 'Container' as it's base.`,
+];
+
+export { TITLE, DESCRIPTION, DO, DONT };