+++ packages/react/src/stories/StoryTitle.tsx
@@ -1,3 +1,4 @@
+import Container from '../components/Container';
 import Typography from '../components/Typography';
 
 type ChildrenProps = {
@@ -5,9 +6,13 @@ type ChildrenProps = {
 };
 
 export const Overview = ({ children }: ChildrenProps) => (
-  <Typography variant="heading6">{children}</Typography>
+  <Container mb="SpacingBase4">
+    <Typography variant="heading6">{children}</Typography>
+  </Container>
 );
 
 export const Regular = ({ children }: ChildrenProps) => (
-  <Typography variant="subtitle1">{children}</Typography>
+  <Container mb="SpacingBase8">
+    <Typography variant="subtitle1">{children}</Typography>
+  </Container>
 );